+++ CHANGELOG.md
@@ -3,6 +3,7 @@ UNRELEASED
 
 * [ [#267](https://github.com/digitalfabrik/lunes-cms/issues/267) ] Add professions in the vocabulary list
 * [ [#497](https://github.com/digitalfabrik/lunes-cms/issues/497) ] Mark words that have an image with icon in training sets view
+* [ [#458](https://github.com/digitalfabrik/lunes-cms/issues/458) ] Add sponsor field to discipline
 
 
 2023.7.0